requirements.txt
Copystreamlit==1.29.0
pandas==2.1.0
plotly==5.18.0
numpy==1.24.3
openpyxl==3.1.2